Bring int/uint in line with RFC 544

+++ src/alsa.rs
@@ -16,7 +16,7 @@ macro_rules! alsa_ok(
         {
             let err = $e;
             if err < 0 {
-                return Err(err as int)
+                return Err(err as isize)
             }
             err
         }
@@ -33,7 +33,7 @@ pub struct Open { #[allow(dead_code)] no_constr: () }
 
 #[allow(missing_copy_implementations)]
 pub struct Prepared {
-    channels: uint,
+    channels: usize,
     sample_fmt: Format
 }
 
@@ -100,7 +100,7 @@ impl Format {
         }
     }
 
-    fn size(self) -> uint {
+    fn size(self) -> usize {
         use std::mem::size_of;
         match self {
             Format::Unsigned8 => 1,
@@ -111,7 +111,7 @@ impl Format {
 }
 
 impl PCM<Open> {
-    pub fn open(name: &str, stream: Stream, mode: Mode) -> Result<PCM<Open>, int> {
+    pub fn open(name: &str, stream: Stream, mode: Mode) -> Result<PCM<Open>, isize> {
         let mut pcm = PCM {
             i: ptr::null_mut(),
             data: Open { no_constr: () }
@@ -129,14 +129,14 @@ impl PCM<Open> {
 }
 
 impl PCM<Open> {
-    pub fn set_parameters(self, format: Format, access: Access, channels: uint, rate: uint)
-        -> Result<PCM<Prepared>, (PCM<Open>, int)>
+    pub fn set_parameters(self, format: Format, access: Access, channels: usize, rate: usize)
+        -> Result<PCM<Prepared>, (PCM<Open>, isize)>
     {
         unsafe {
             let err = ffi::snd_pcm_set_params(self.i, format.to_ffi(), access.to_ffi(),
                                               channels as u32, rate as u32, 1i32, 500000u32);
             if err < 0 {
-                return Err((self, err as int))
+                return Err((self, err as isize))
             }
         }
 
@@ -154,7 +154,7 @@ impl PCM<Open> {
 }
 
 impl PCM<Prepared> {
-    pub fn write_interleaved<T: Copy>(&mut self, buffer: &[T]) -> Result<uint, int> {
+    pub fn write_interleaved<T: Copy>(&mut self, buffer: &[T]) -> Result<usize, isize> {
         let channels = self.data.channels;
 
         assert_eq!(buffer.len() % channels, 0);
@@ -165,6 +165,6 @@ impl PCM<Prepared> {
                                          buffer.len() as u64 / channels as u64))
         };
 
-        Ok(n_written as uint)
+        Ok(n_written as usize)
     }
 }
